Revert "Issue #SB-6391 feat: Content import offline functionality (#1790)"
This reverts commit 34e6c0b6e96431a8cd9603109443c1aaed7cec4f.
