+++ src/app/client/angular.json
@@ -174,7 +174,6 @@
               ]
             },
             "scripts": [
-              "src/assets/libs/iziToast/iziToast.min.js",
               "node_modules/jquery/dist/jquery.min.js",
               "node_modules/slick-carousel/slick/slick.min.js",
               "node_modules/@project-sunbird/telemetry-sdk/index.js",
