+++ src/app/client/src/app/framework.config.ts
@@ -1,13 +1,6 @@
-import { OfflineModule } from './modules/offline';
-import { ImportContentPlugin} from './modules/offline/plugin';
 
 export const WebExtensionsConfig = {
-  plugins: [{
-    'id': 'content-import',
-    'ver': '1.0.0',
-    'module': OfflineModule,
-    'main': ImportContentPlugin
-  }]
+  plugins: []
 };
 
-export const PluginModules = [ OfflineModule ];
+export const PluginModules = [ ];
