+++ src/app/client/src/app/modules/core/components/main-menu/main-menu.component.html
@@ -12,8 +12,8 @@
 			{{resourceService?.frmelmnts?.tab?.courses}}
 		</a>
 		<a class="item" [class.active]="router.isActive('/workspace')" appTelemetryInteract
-			[telemetryInteractEdata]="workspaceMenuIntractEdata" *ngIf='permissionService.permissionAvailable'
-			appPermission [permission]='workSpaceRole' [routerLink]="[navigateToWorkspace()]" tabindex="0">
+			[telemetryInteractEdata]="workspaceMenuIntractEdata" *ngIf='permissionService.permissionAvailable' appPermission
+			[permission]='workSpaceRole' [routerLink]="[navigateToWorkspace()]" tabindex="0">
 			{{resourceService?.frmelmnts?.tab?.workspace}}
 		</a>
 		<a class="item" appTelemetryInteract [telemetryInteractEdata]="helpMenuIntractEdata"
@@ -26,14 +26,14 @@
 
 	<!--Not Logged in user-->
 	<ng-container *ngIf="!userService.loggedIn && !isOffline">
-		<button class="ui blue button" *ngIf="router.isActive('', true) && exploreButtonVisibility==='true'"
-			tabindex="0" [routerLink]="'/explore'">
+		<button class="ui blue button" *ngIf="router.isActive('', true) && exploreButtonVisibility==='true'" tabindex="0"
+			[routerLink]="'/explore'">
 			{{resourceService.frmelmnts?.lbl?.explore}} {{resourceService.instance}}
 		</button>
 		<a class="item"
 			[class.active]="router.isActive('/search/explore') || router.isActive(slug ? slug + '/explore' : '/explore')"
-			*ngIf="!router.isActive('', true)" [routerLink]="[slug ? slug + '/explore' : '/explore']"
-			appTelemetryInteract [telemetryInteractEdata]="libraryMenuIntractEdata">
+			*ngIf="!router.isActive('', true)" [routerLink]="[slug ? slug + '/explore' : '/explore']" appTelemetryInteract
+			[telemetryInteractEdata]="libraryMenuIntractEdata">
 			{{resourceService?.frmelmnts?.tab?.resources}}
 		</a>
 		<a class="item"
@@ -49,22 +49,9 @@
 		</a>
 	</ng-container>
 	<!--/Not Logged in user-->
-	<!-- offline container-->
-	<ng-container *ngIf="isOffline">
-		<a class="item" [routerLink]="'/'" appTelemetryInteract [telemetryInteractEdata]="myLibraryMenuInteractEdata">
-			{{resourceService.frmelmnts?.btn?.myLibrary}}
-		</a>
-		<a class="item" [routerLink]="'/'" appTelemetryInteract [telemetryInteractEdata]="browseEdata">
-			{{resourceService.frmelmnts?.btn?.browse}}
-		</a>
-		<!-- <a class="item" [routerLink]="'/'">
-			{{resourceService.frmelmnts?.btn?.settings}}
-		</a> -->
-	</ng-container>
-	<!-- offline container-->
 
 </div>
-<!--Computer-->
+<!--/Computer-->
 
 <!--Mobile-->
 <div class="mobile only">
@@ -114,21 +101,16 @@
 			<a class="item" (click)="showSideBar()" *ngIf="router.isActive('', true)" routerLink="/explore">
 				{{resourceService.frmelmnts?.lbl?.explore}} {{resourceService.instance}}
 			</a>
-			<a class="item"
-				[class.active]="router.isActive('/search/explore') || router.isActive(slug ? slug + '/explore' : '/explore')"
-				[routerLink]="[slug ? slug + '/explore' : '/explore']" (click)="showSideBar()"
-				*ngIf="!router.isActive('', true)">
+			<a class="item" [class.active]="router.isActive('/search/explore') || router.isActive(slug ? slug + '/explore' : '/explore')"
+			[routerLink]="[slug ? slug + '/explore' : '/explore']" (click)="showSideBar()" *ngIf="!router.isActive('', true)">
 				{{resourceService?.frmelmnts?.tab?.resources}}
 			</a>
-			<a class="item"
-				[class.active]="router.isActive('/search/explore-course') || router.isActive(slug ? slug + '/explore-course' : '/explore-course')"
-				[routerLink]="[slug ? slug + '/explore-course' : '/explore-course']" (click)="showSideBar()"
-				*ngIf="!router.isActive('', true)">
+			<a class="item" [class.active]="router.isActive('/search/explore-course') || router.isActive(slug ? slug + '/explore-course' : '/explore-course')"
+			[routerLink]="[slug ? slug + '/explore-course' : '/explore-course']" (click)="showSideBar()" *ngIf="!router.isActive('', true)">
 				{{resourceService?.frmelmnts?.tab?.courses}}
 			</a>
 		</ng-container>
 		<!--/Not Logged in user-->
-
 
 	</div>
 </div>
