+++ src/app/client/src/app/modules/core/components/main-menu/main-menu.component.ts
@@ -49,8 +49,6 @@ export class MainMenuComponent implements OnInit {
   homeMenuIntractEdata: IInteractEventEdata;
   learnMenuIntractEdata: IInteractEventEdata;
   libraryMenuIntractEdata: IInteractEventEdata;
-  myLibraryMenuInteractEdata: IInteractEventEdata;
-  browseEdata: IInteractEventEdata;
   workspaceMenuIntractEdata: IInteractEventEdata;
   helpMenuIntractEdata: IInteractEventEdata;
   exploreRoutingUrl: string;
@@ -99,16 +97,6 @@ export class MainMenuComponent implements OnInit {
       id: 'library-tab',
       type: 'click',
       pageid: 'library'
-    };
-    this.myLibraryMenuInteractEdata = {
-      id: 'myLibrary-tab',
-      type: 'click',
-      pageid: 'library'
-    };
-    this.browseEdata = {
-      id: 'browse-tab',
-      type: 'click',
-      pageid: 'browse'
     };
     this.learnMenuIntractEdata = {
       id: 'learn-tab',
