--- src/app/client/src/app/modules/offline/components/content-import-header/content-import-header.component.scss
@@ -1,33 +1,0 @@
-/* temp */
-.sb-offline-imports {
-    background:white;
-}
-.sb-offline-area {
-    padding:16px 0px;min-height:80px;align-items: center;
-  }
-  .sb-import-icon {
-    height: 56px;
-    width: 56px;
-    border-radius: 50%;
-    background-color: rgba(0,136,64,0.1);
-    text-align: center;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-right: 16px;
-    i {
-      font-size: 32px;
-    }
-  }
-  .sb-import-content {
-    flex: 50%;
-    h4 {
-      color: #024F9D;
-      font-size: 14px;
-      line-height: 20px;
-      margin-bottom: 0px;
-    }
-    p {
-      color: #333333;	font-size: 12px;	line-height: 18px;
-    }
-  }
