--- src/app/client/src/app/modules/offline/components/content-import-header/content-import-header.component.ts
@@ -1,33 +1,0 @@
-import { IInteractEventEdata, IInteractEventObject } from '@sunbird/telemetry';
-import { Component, OnInit } from '@angular/core';
-import { ResourceService } from '@sunbird/shared';
-
-@Component({
-    selector: 'app-content-import-header',
-    templateUrl: './content-import-header.component.html',
-    styleUrls: ['./content-import-header.component.scss']
-})
-export class ContentImportHeaderComponent implements OnInit {
-    showImportModal = false;
-    ContentImportIntractEdata: IInteractEventEdata;
-    public telemetryInteractObject: IInteractEventObject;
-
-    constructor(public resourceService: ResourceService) { }
-
-    ngOnInit() {
-        this.setInteractData();
-    }
-
-    setInteractData() {
-        this.telemetryInteractObject = {
-            id: 'content-import',
-            type: 'content-import',
-            ver: '1.0'
-          };
-        this.ContentImportIntractEdata = {
-            id: 'content-import',
-            type: 'click',
-            pageid: 'library'
-        };
-    }
-}
