+++ src/app/client/src/app/modules/offline/components/content-import/content-import.component.spec.ts
@@ -1,10 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ContentImportComponent } from './content-import.component';
-import { OfflineFileUploaderService } from '../../services';
-import { SuiModalModule } from 'ng2-semantic-ui';
-import { SharedModule } from '@sunbird/shared';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 describe('ContentImportComponent', () => {
   let component: ContentImportComponent;
@@ -12,9 +8,7 @@ describe('ContentImportComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ SuiModalModule, SharedModule.forRoot(), HttpClientTestingModule ],
-      declarations: [ContentImportComponent],
-      providers: [ OfflineFileUploaderService ]
+      declarations: [ContentImportComponent]
     })
       .compileComponents();
   }));
@@ -25,4 +19,7 @@ describe('ContentImportComponent', () => {
     fixture.detectChanges();
   });
 
+  xit('should create', () => {
+    expect(component).toBeTruthy();
+  });
 });
