+++ src/app/client/src/app/modules/offline/components/index.ts
@@ -1,2 +1,1 @@
 export * from './content-import/content-import.component';
-export * from './content-import-header/content-import-header.component';
