+++ src/app/client/src/app/modules/offline/index.ts
@@ -1,4 +1,2 @@
 export * from './offline.module';
 export * from './components';
-export * from './services';
-
