+++ src/app/client/src/app/modules/offline/offline.module.ts
@@ -1,30 +1,19 @@
-import { OfflineFileUploaderService } from './services';
-import { SuiModalModule } from 'ng2-semantic-ui';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CoreModule } from '@sunbird/core';
+import { PublicPlayerService } from '@sunbird/public';
 import { SharedModule } from '@sunbird/shared';
 import { OfflineRoutingModule } from './offline-routing.module';
 import { DeviceDetectorService } from 'ngx-device-detector';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ContentImportComponent, ContentImportHeaderComponent } from './components';
-
-
+import { ContentImportComponent } from './components/content-import/content-import.component';
 @NgModule({
   imports: [
-    BrowserAnimationsModule,
     CommonModule,
     CoreModule,
     SharedModule,
-    OfflineRoutingModule,
-    SuiModalModule,
+    OfflineRoutingModule
   ],
-  providers: [DeviceDetectorService, OfflineFileUploaderService],
-  declarations: [
-    ContentImportComponent, ContentImportHeaderComponent
-  ],
-  entryComponents: [
-    ContentImportHeaderComponent
-  ]
+  providers: [PublicPlayerService, DeviceDetectorService],
+  declarations: [ContentImportComponent]
 })
 export class OfflineModule { }
