--- src/app/client/src/app/modules/offline/services/index.ts
@@ -1,1 +1,0 @@
-export * from './offline-file-uploader-service/offline-file-uploader.service';
