--- src/app/client/src/app/modules/offline/services/offline-file-uploader-service/offline-file-uploader.service.spec.ts
@@ -1,18 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { OfflineFileUploaderService } from './offline-file-uploader.service';
-import { ToasterService } from '@sunbird/shared';
-
-describe('OfflineFileUploaderService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [],
-      providers: [ToasterService]
-    });
-  });
-
-  it('should be created', () => {
-    const service: OfflineFileUploaderService = TestBed.get(OfflineFileUploaderService);
-    expect(service).toBeTruthy();
-  });
-});
