+++ src/app/resourcebundles/data/consumption/en.properties
@@ -39,12 +39,7 @@ frmelmnts.btn.create=Create
 frmelmnts.btn.copyLink = Copy link
 frmelmnts.btn.resume=Resume
 frmelmnts.btn.cancelCapitalize=Cancel
-frmelmnts.btn.myLibrary=My Library
-frmelmnts.btn.browse=Browse
-frmelmnts.btn.settings=Settings
-frmelmnts.btn.contentImport=Upload Files
-frmelmnts.btn.processingFiles=Processing files
-frmelmnts.btn.selectContentFiles=Select file(s)
+
 
 
 # Labels
@@ -426,18 +421,6 @@ frmelmnts.scttl.blkuser=Block User
 frmelmnts.scttl.contributions=Contribution(s)
 frmelmnts.lbl.plslgn=This session has expired. Please login again to continue using ${instance}.
 frmelmnts.lbl.sesnexrd=Session Expired
-frmelmnts.lbl.uploadEcarFromPd=Upload .ecar files from a pen drive to import content
-frmelmnts.lbl.click=Click
-frmelmnts.lbl.choseFromPd=to choose files from your pen drive.
-frmelmnts.lbl.open=Open
-frmelmnts.lbl.inWindows=in the Windows dialogue box.
-frmelmnts.lbl.afterSuccessfulImport=After you get a message that the files are successfully\ imported, click
-frmelmnts.lbl.afterUnsuccessfulImport=If the files are not imported successfully, click
-frmelmnts.lbl.retry=Retry
-frmelmnts.lbl.retryImport=to try and upload the files once again, or click
-frmelmnts.lbl.haltOperation=to halt the operation.
-frmelmnts.lbl.contentsUploaded=Contents are being uploaded
-
 # Dropdowns
 frmelmnts.drpdn.male=Male
 frmelmnts.drpdn.female=Female
